better error handling

+++ beverages/BeverageController.go
@@ -28,14 +28,22 @@ func (controller *BeverageController) GetBeverages(ctx *gin.Context) {
 	if err != nil {
 		errResp, _ := restapi.NewErrorResponse("Couldnt marshal the beverage array").Marshal()
 		fmt.Fprint(ctx.Writer, string(errResp))
-	} else {
-		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Abort()
+		return
 	}
+	fmt.Fprint(ctx.Writer, string(response))
+	ctx.Next()
 }
 
 //GetBeverage responds with the beverage identified by beverage/:id
 func (controller *BeverageController) GetBeverage(ctx *gin.Context) {
-	ID, _ := ctx.GetQuery("id")
+	ID, ok := ctx.GetQuery("id")
+	if !ok {
+		errResp, _ := restapi.NewErrorResponse("No ID given").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
 
 	bev, ok := controller.service.GetBeverage(ID)
 	if ok {
@@ -43,53 +51,102 @@ func (controller *BeverageController) GetBeverage(ctx *gin.Context) {
 		if err != nil {
 			errResp, _ := restapi.NewErrorResponse("Couldnt marshal the beverage object").Marshal()
 			fmt.Fprint(ctx.Writer, string(errResp))
-		} else {
-			fmt.Fprint(ctx.Writer, string(response))
+			ctx.Abort()
+			return
 		}
+		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Next()
 	} else {
 		response, _ := restapi.NewErrorResponse("").Marshal()
 		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Abort()
+		return
 	}
 }
 
 //NewBeverage creates a new beverage with the given form-values "value" and "name" and returns it
 func (controller *BeverageController) NewBeverage(ctx *gin.Context) {
-	nv, _ := strconv.Atoi(ctx.PostForm("value"))
-	nb, _ := controller.service.NewBeverage(ctx.PostForm("name"), nv)
+	nv, err := strconv.Atoi(ctx.PostForm("value"))
+	if err != nil {
+		errResp, _ := restapi.NewErrorResponse("Invalid value").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
+
+	nb, ok := controller.service.NewBeverage(ctx.PostForm("name"), nv)
+	if !ok {
+		errResp, _ := restapi.NewErrorResponse("Couldnt save new beverage").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
 	response, err := restapi.NewOkResponse(nb).Marshal()
 	if err != nil {
 		errResp, _ := restapi.NewErrorResponse("Couldnt marshal the beverage object").Marshal()
 		fmt.Fprint(ctx.Writer, string(errResp))
-	} else {
-		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Abort()
+		return
 	}
+	fmt.Fprint(ctx.Writer, string(response))
+	ctx.Next()
 }
 
 //UpdateBeverage updates the beverage identified by /beverage/:id with the given form-values "value" and "name" and returns it
 func (controller *BeverageController) UpdateBeverage(ctx *gin.Context) {
-	ID, _ := ctx.GetQuery("id")
+	ID, ok := ctx.GetQuery("id")
+	if !ok {
+		errResp, _ := restapi.NewErrorResponse("No ID given").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
 
-	nv, _ := strconv.Atoi(ctx.PostForm("value"))
+	nv, err := strconv.Atoi(ctx.PostForm("value"))
+	if err != nil {
+		errResp, _ := restapi.NewErrorResponse("Invalid value").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
 	nn := ctx.PostForm("name")
-	nb, _ := controller.service.UpdateBeverage(ID, nn, nv)
+	nb, ok := controller.service.UpdateBeverage(ID, nn, nv)
+	if !ok {
+		errResp, _ := restapi.NewErrorResponse("Couldnt update beverage").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
+
 	response, err := restapi.NewOkResponse(nb).Marshal()
 	if err != nil {
 		errResp, _ := restapi.NewErrorResponse("Couldnt marshal the beverage object").Marshal()
 		fmt.Fprint(ctx.Writer, string(errResp))
-	} else {
-		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Abort()
+		return
 	}
+	fmt.Fprint(ctx.Writer, string(response))
+	ctx.Next()
 }
 
 //DeleteBeverage deletes the beverage identified by /beverage/:id and responds with a YEP/NOPE
 func (controller *BeverageController) DeleteBeverage(ctx *gin.Context) {
-	ID, _ := ctx.GetQuery("id")
+	ID, ok := ctx.GetQuery("id")
+	if !ok {
+		errResp, _ := restapi.NewErrorResponse("No ID given").Marshal()
+		fmt.Fprint(ctx.Writer, string(errResp))
+		ctx.Abort()
+		return
+	}
 
 	if controller.service.DeleteBeverage(ID) {
 		response, _ := restapi.NewOkResponse("").Marshal()
 		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Next()
 	} else {
 		response, _ := restapi.NewErrorResponse("").Marshal()
 		fmt.Fprint(ctx.Writer, string(response))
+		ctx.Abort()
+		return
 	}
 }
